fixed warnings from Linux compilation

+++ sample.c
@@ -470,7 +470,7 @@ int offset_compare(const void *off1, const void *off2)
 {
 #ifdef DEBUG
     fprintf(stderr, "Debug: offset_compare()\n");
-    fprintf(stderr, "Debug: Comparing: %012lld and %012lld\n", *(off_t *)off1, *(off_t *)off2);
+    fprintf(stderr, "Debug: Comparing: %012lld and %012lld\n", (long long int) *(off_t *)off1, (long long int) *(off_t *)off2);
 #endif
     int off_diff = (*(off_t *)off1) - (*(off_t *)off2);
     return (off_diff > 0) ? 1 : -1;
